fix(kit-demo): stop Gestures/Scroll tabs jumping to top; split GesturesSection into cards
Keep the page header + top padding when switching to the non-scrolling tabs so the tab bar no longer leaps upward; only the outer element swaps to a non-scrolling container so the inner gesture/scroll surface still owns the touch stream. Extract SwipeAction and Sortable into self-contained subcomponents.
